+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 2.4.0
 
-- **[add] Chained methods now work on Form openers too**
+- **[add] Form openers are now objects too and accept chained methods**
 - [add] Add `unchecked_value` option to decide what value unchecked checkboxes have in the POST array
 - [add] Allow booleans to be passed to Checkable->check() on single items
 - [mod] Disable `push_checkbox` option by default